+++ backend/connector-integration/src/connectors.rs
@@ -194,5 +194,8 @@ pub use self::payme::Payme;
 pub mod revolut;
 pub use self::revolut::Revolut;
 
+pub mod gigadat;
+pub use self::gigadat::Gigadat;
+
 pub mod loonio;
 pub use self::loonio::Loonio;